+++ docs/Changelog.rst
@@ -1,7 +1,8 @@
 2.7.0 (unreleased)
 ------------------
 
-- ##172 Support AllowManualEntry for select result options
+- #172 Support AllowManualEntry for select result options
+- #173 Add global "Fetch transitions on select" listing setting
 - #171 Fix TypeError in SearchableSelect when option values are non-string
 - #170 Fix UnicodeEncodeError when column filter values contain non-ASCII
 - #169 Allow per-cell field type override and display value in HiddenField